week_5/flipkart-recommender/requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
chromadb
sentence-transformers
pandas
flask
prometheus_client
python-dotenv
